+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 * Add support for lazy loaded page in `admin-panel` template
 * Add superuser member for `admin-panel` template
+* API endpoint for get members (only for superuser) in `admin-panel` template
 
 ## 4.0.3
 * Fix login in `admin-panel` template